Fixed a help bug that prevented the correct scrolling to the requested page reference
